Added Introducing The Graph section
Proposing a new sub-section called Introducing The Graph to the Advocates Program main section. This new page enables to remove this language from the actual Advocate application/questionnaire and instead reference it there as a link to GitBook. It aims to streamline the application processing time and thereby improving the user experience for Advocates program applicants, especially to those already familiar with the Graph ecosystem and protocol. Actual language for the page will be proposed in a separate PR