Introducing GOC main section
GOC-0001 was approved and adopted by the DAO on May 4th, 2022 in DAOHaus. This new GOC main section aims to create a version controlled place for GOCs to be drafted and eventually be transparently published once individual GOCs are approved by the DAO. GOC-0002 through GOC-0016 have been added as placeholder pages. They are planned future GOCs that the DAO is looking to develop.